+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -131,7 +131,8 @@ export class PortfolioController {
           portfolioPosition.investment / totalInvestment;
         portfolioPosition.netPerformance = null;
         portfolioPosition.quantity = null;
-        portfolioPosition.value = portfolioPosition.value / totalValue;
+        portfolioPosition.valueInPercentage =
+          portfolioPosition.value / totalValue;
       }
 
       for (const [name, { current, original }] of Object.entries(accounts)) {
@@ -322,7 +323,7 @@ export class PortfolioController {
             totalInvestment: new Big(totalInvestment)
               .div(performanceInformation.performance.totalInvestment)
               .toNumber(),
-            value: new Big(value)
+            valueInPercentage: new Big(value)
               .div(performanceInformation.performance.currentValue)
               .toNumber()
           };
@@ -356,6 +357,7 @@ export class PortfolioController {
 
   @Get('positions')
   @UseGuards(AuthGuard('jwt'))
+  @UseInterceptors(RedactValuesInResponseInterceptor)
   @UseInterceptors(TransformDataSourceInResponseInterceptor)
   public async getPositions(
     @Headers('impersonation-id') impersonationId: string,
@@ -370,27 +372,11 @@ export class PortfolioController {
       filterByTags
     });
 
-    const result = await this.portfolioService.getPositions({
+    return this.portfolioService.getPositions({
       dateRange,
       filters,
       impersonationId
     });
-
-    if (
-      impersonationId ||
-      this.userService.isRestrictedView(this.request.user)
-    ) {
-      result.positions = result.positions.map((position) => {
-        return nullifyValuesInObject(position, [
-          'grossPerformance',
-          'investment',
-          'netPerformance',
-          'quantity'
-        ]);
-      });
-    }
-
-    return result;
   }
 
   @Get('public/:accessId')
@@ -452,7 +438,7 @@ export class PortfolioController {
         sectors: hasDetails ? portfolioPosition.sectors : [],
         symbol: portfolioPosition.symbol,
         url: portfolioPosition.url,
-        value: portfolioPosition.value / totalValue
+        valueInPercentage: portfolioPosition.value / totalValue
       };
     }
 
@@ -460,6 +446,7 @@ export class PortfolioController {
   }
 
   @Get('position/:dataSource/:symbol')
+  @UseInterceptors(RedactValuesInResponseInterceptor)
   @UseInterceptors(TransformDataSourceInRequestInterceptor)
   @UseInterceptors(TransformDataSourceInResponseInterceptor)
   @UseGuards(AuthGuard('jwt'))
@@ -468,27 +455,13 @@ export class PortfolioController {
     @Param('dataSource') dataSource,
     @Param('symbol') symbol
   ): Promise<PortfolioPositionDetail> {
-    let position = await this.portfolioService.getPosition(
+    const position = await this.portfolioService.getPosition(
       dataSource,
       impersonationId,
       symbol
     );
 
     if (position) {
-      if (
-        impersonationId ||
-        this.userService.isRestrictedView(this.request.user)
-      ) {
-        position = nullifyValuesInObject(position, [
-          'grossPerformance',
-          'investment',
-          'netPerformance',
-          'orders',
-          'quantity',
-          'value'
-        ]);
-      }
-
       return position;
     }