+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -24,7 +24,7 @@ import {
   MAX_CHART_ITEMS,
   UNKNOWN_KEY
 } from '@ghostfolio/common/config';
-import { DATE_FORMAT, parseDate } from '@ghostfolio/common/helper';
+import { DATE_FORMAT, getSum, parseDate } from '@ghostfolio/common/helper';
 import {
   Accounts,
   EnhancedSymbolProfile,
@@ -573,7 +573,6 @@ export class PortfolioService {
       const cashPositions = await this.getCashPositions({
         cashDetails,
         userCurrency,
-        investment: totalInvestmentInBaseCurrency,
         value: filteredValueInBaseCurrency
       });
 
@@ -599,7 +598,6 @@ export class PortfolioService {
       const cashPositions = await this.getCashPositions({
         cashDetails,
         userCurrency,
-        investment: totalInvestmentInBaseCurrency,
         value: filteredValueInBaseCurrency
       });
 
@@ -680,6 +678,8 @@ export class PortfolioService {
       return {
         tags,
         averagePrice: undefined,
+        dividendInBaseCurrency: undefined,
+        feeInBaseCurrency: undefined,
         firstBuyDate: undefined,
         grossPerformance: undefined,
         grossPerformancePercent: undefined,
@@ -746,11 +746,22 @@ export class PortfolioService {
         averagePrice,
         currency,
         dataSource,
+        fee,
         firstBuyDate,
         marketPrice,
         quantity,
         transactionCount
       } = position;
+
+      const dividendInBaseCurrency = getSum(
+        orders
+          .filter(({ type }) => {
+            return type === 'DIVIDEND';
+          })
+          .map(({ valueInBaseCurrency }) => {
+            return new Big(valueInBaseCurrency);
+          })
+      );
 
       // Convert investment, gross and net performance to currency of user
       const investment = this.exchangeRateDataService.toCurrency(
@@ -785,8 +796,8 @@ export class PortfolioService {
         historicalDataArray.push({
           averagePrice: orders[0].unitPrice,
           date: firstBuyDate,
-          quantity: orders[0].quantity,
-          value: orders[0].unitPrice
+          marketPrice: orders[0].unitPrice,
+          quantity: orders[0].quantity
         });
       }
 
@@ -815,9 +826,9 @@ export class PortfolioService {
 
           historicalDataArray.push({
             date,
+            marketPrice,
             averagePrice: currentAveragePrice,
-            quantity: currentQuantity,
-            value: marketPrice
+            quantity: currentQuantity
           });
 
           maxPrice = Math.max(marketPrice ?? 0, maxPrice);
@@ -838,6 +849,12 @@ export class PortfolioService {
         tags,
         transactionCount,
         averagePrice: averagePrice.toNumber(),
+        dividendInBaseCurrency: dividendInBaseCurrency.toNumber(),
+        feeInBaseCurrency: this.exchangeRateDataService.toCurrency(
+          fee.toNumber(),
+          SymbolProfile.currency,
+          userCurrency
+        ),
         grossPerformancePercent:
           position.grossPerformancePercentage?.toNumber(),
         historicalData: historicalDataArray,
@@ -894,6 +911,8 @@ export class PortfolioService {
         SymbolProfile,
         tags,
         averagePrice: 0,
+        dividendInBaseCurrency: 0,
+        feeInBaseCurrency: 0,
         firstBuyDate: undefined,
         grossPerformance: undefined,
         grossPerformancePercent: undefined,
@@ -1039,29 +1058,21 @@ export class PortfolioService {
 
     const portfolioStart = parseDate(transactionPoints[0].date);
     const startDate = this.getStartDate(dateRange, portfolioStart);
-    const currentPositions = await portfolioCalculator.getCurrentPositions(
-      startDate
-    );
+    const {
+      currentValue,
+      errors,
+      grossPerformance,
+      grossPerformancePercentage,
+      hasErrors,
+      netPerformance,
+      netPerformancePercentage,
+      totalInvestment
+    } = await portfolioCalculator.getCurrentPositions(startDate);
 
-    const hasErrors = currentPositions.hasErrors;
-    const currentValue = currentPositions.currentValue.toNumber();
-    const currentGrossPerformance = currentPositions.grossPerformance;
-    const currentGrossPerformancePercent =
-      currentPositions.grossPerformancePercentage;
-    let currentNetPerformance = currentPositions.netPerformance;
-    let currentNetPerformancePercent =
-      currentPositions.netPerformancePercentage;
-    const totalInvestment = currentPositions.totalInvestment;
-
-    // if (currentGrossPerformance.mul(currentGrossPerformancePercent).lt(0)) {
-    //   // If algebraic sign is different, harmonize it
-    //   currentGrossPerformancePercent = currentGrossPerformancePercent.mul(-1);
-    // }
-
-    // if (currentNetPerformance.mul(currentNetPerformancePercent).lt(0)) {
-    //   // If algebraic sign is different, harmonize it
-    //   currentNetPerformancePercent = currentNetPerformancePercent.mul(-1);
-    // }
+    const currentGrossPerformance = grossPerformance;
+    const currentGrossPerformancePercent = grossPerformancePercentage;
+    let currentNetPerformance = netPerformance;
+    let currentNetPerformancePercent = netPerformancePercentage;
 
     const historicalDataContainer = await this.getChart({
       dateRange,
@@ -1083,28 +1094,28 @@ export class PortfolioService {
     }
 
     return {
+      errors,
+      hasErrors,
       chart: historicalDataContainer.items.map(
         ({
           date,
-          netPerformance,
+          netPerformance: netPerformanceOfItem,
           netPerformanceInPercentage,
-          totalInvestment,
+          totalInvestment: totalInvestmentOfItem,
           value
         }) => {
           return {
             date,
-            netPerformance,
             netPerformanceInPercentage,
-            totalInvestment,
-            value
+            value,
+            netPerformance: netPerformanceOfItem,
+            totalInvestment: totalInvestmentOfItem
           };
         }
       ),
-      errors: currentPositions.errors,
       firstOrderDate: parseDate(historicalDataContainer.items[0]?.date),
-      hasErrors: currentPositions.hasErrors || hasErrors,
       performance: {
-        currentValue,
+        currentValue: currentValue.toNumber(),
         currentGrossPerformance: currentGrossPerformance.toNumber(),
         currentGrossPerformancePercent:
           currentGrossPerformancePercent.toNumber(),
@@ -1217,12 +1228,10 @@ export class PortfolioService {
 
   private async getCashPositions({
     cashDetails,
-    investment,
     userCurrency,
     value
   }: {
     cashDetails: CashDetails;
-    investment: Big;
     userCurrency: string;
     value: Big;
   }) {
@@ -1700,6 +1709,14 @@ export class PortfolioService {
     userId: string;
     withExcludedAccounts?: boolean;
   }) {
+    const ordersOfTypeItem = await this.orderService.getOrders({
+      filters,
+      userCurrency,
+      userId,
+      withExcludedAccounts,
+      types: ['ITEM']
+    });
+
     const accounts: PortfolioDetails['accounts'] = {};
 
     let currentAccounts: (Account & {
@@ -1730,9 +1747,17 @@ export class PortfolioService {
     });
 
     for (const account of currentAccounts) {
-      const ordersByAccount = orders.filter(({ accountId }) => {
+      let ordersByAccount = orders.filter(({ accountId }) => {
         return accountId === account.id;
       });
+
+      const ordersOfTypeItemByAccount = ordersOfTypeItem.filter(
+        ({ accountId }) => {
+          return accountId === account.id;
+        }
+      );
+
+      ordersByAccount = ordersByAccount.concat(ordersOfTypeItemByAccount);
 
       accounts[account.id] = {
         balance: account.balance,
@@ -1753,7 +1778,9 @@ export class PortfolioService {
       for (const order of ordersByAccount) {
         let currentValueOfSymbolInBaseCurrency =
           order.quantity *
-            portfolioItemsNow[order.SymbolProfile.symbol]?.marketPrice ?? 0;
+          (portfolioItemsNow[order.SymbolProfile.symbol]?.marketPrice ??
+            order.unitPrice ??
+            0);
         let originalValueOfSymbolInBaseCurrency =
           this.exchangeRateDataService.toCurrency(
             order.quantity * order.unitPrice,