+++ apps/client/src/app/services/user/user.service.ts
@@ -1,10 +1,16 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { MatLegacyDialog as MatDialog } from '@angular/material/legacy-dialog';
 import { ObservableStore } from '@codewithdan/observable-store';
+import { SubscriptionInterstitialDialogParams } from '@ghostfolio/client/components/subscription-interstitial-dialog/interfaces/interfaces';
+import { SubscriptionInterstitialDialog } from '@ghostfolio/client/components/subscription-interstitial-dialog/subscription-interstitial-dialog.component';
 import { User } from '@ghostfolio/common/interfaces';
-import { of } from 'rxjs';
+import { hasPermission, permissions } from '@ghostfolio/common/permissions';
+import { timezoneCitiesToCountries } from '@ghostfolio/common/timezone-cities-to-countries';
+import { DeviceDetectorService } from 'ngx-device-detector';
+import { Subject, of } from 'rxjs';
 import { throwError } from 'rxjs';
-import { catchError, map } from 'rxjs/operators';
+import { catchError, map, takeUntil } from 'rxjs/operators';
 
 import { UserStoreActions } from './user-store.actions';
 import { UserStoreState } from './user-store.state';
@@ -13,10 +19,19 @@ import { UserStoreState } from './user-store.state';
   providedIn: 'root'
 })
 export class UserService extends ObservableStore<UserStoreState> {
-  public constructor(private http: HttpClient) {
+  private deviceType: string;
+  private unsubscribeSubject = new Subject<void>();
+
+  public constructor(
+    private deviceService: DeviceDetectorService,
+    private dialog: MatDialog,
+    private http: HttpClient
+  ) {
     super({ trackStateHistory: true });
 
     this.setState({ user: undefined }, UserStoreActions.Initialize);
+
+    this.deviceType = this.deviceService.getDeviceInfo().deviceType;
   }
 
   public get(force = false) {
@@ -31,6 +46,20 @@ export class UserService extends ObservableStore<UserStoreState> {
     }
   }
 
+  public getCountry() {
+    let country: string;
+
+    if (Intl) {
+      const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+      const timeZoneArray = timeZone.split('/');
+      const city = timeZoneArray[timeZoneArray.length - 1];
+
+      country = timezoneCitiesToCountries[city];
+    }
+
+    return country;
+  }
+
   public remove() {
     this.setState({ user: null }, UserStoreActions.RemoveUser);
   }
@@ -39,6 +68,26 @@ export class UserService extends ObservableStore<UserStoreState> {
     return this.http.get<User>('/api/v1/user').pipe(
       map((user) => {
         this.setState({ user }, UserStoreActions.GetUser);
+
+        if (
+          hasPermission(
+            user.permissions,
+            permissions.enableSubscriptionInterstitial
+          )
+        ) {
+          const dialogRef = this.dialog.open(SubscriptionInterstitialDialog, {
+            autoFocus: false,
+            data: <SubscriptionInterstitialDialogParams>{},
+            height: this.deviceType === 'mobile' ? '97.5vh' : '80vh',
+            width: this.deviceType === 'mobile' ? '100vw' : '50rem'
+          });
+
+          dialogRef
+            .afterClosed()
+            .pipe(takeUntil(this.unsubscribeSubject))
+            .subscribe(() => {});
+        }
+
         return user;
       }),
       catchError(this.handleError)